Implement ToRichIr for Mir
