+++ compiler/frontend/src/mir_optimize/module_folding.rs
@@ -30,7 +30,8 @@
 use crate::{
     mir::{Expression, Id, Mir},
     mir_optimize::OptimizeMir,
-    module::UsePath,
+    module::{Module, UsePath},
+    rich_ir::ToRichIr,
     tracing::TracingConfig,
 };
 use rustc_hash::FxHashMap;
@@ -54,7 +55,10 @@ impl Mir {
                     return; // TODO: Replace with a panic.
                 };
                 let Ok(module_to_import) = path.resolve_relative_to(current_module.clone()) else {
-                    warn!("`use` called with a path that doesn't refer to a module: `\"{path:?}\"` relative to {current_module}.");
+                    warn!(
+                        "`use` called with a path that doesn't refer to a module: `\"{path:?}\"` relative to {}.",
+                        <Module as ToRichIr<Module>>::to_rich_ir(current_module),
+                    );
                     return; // TODO: Replace with a panic.
                 };
 
@@ -63,7 +67,10 @@ impl Mir {
                     tracing.for_child_module(),
                 );
                 let Some(mir) = mir else {
-                    warn!("Module {module_to_import} not found.");
+                    warn!(
+                        "Module {} not found.",
+                        <Module as ToRichIr<Module>>::to_rich_ir(&module_to_import),
+                    );
                     return; // TODO: Replace with a panic.
                 };
                 let mir = (*mir).clone();
