+++ compiler/language_server/src/semantic_tokens.rs
@@ -12,6 +12,7 @@ use crate::utils::range_to_lsp_range_raw;
 
 #[derive(Debug, EnumIter, Hash, PartialEq, Eq, Clone, Copy)]
 pub enum SemanticTokenType {
+    Module,
     Parameter,
     Variable,
     Symbol,
@@ -31,6 +32,7 @@ lazy_static! {
 impl SemanticTokenType {
     pub fn as_lsp(&self) -> lsp_types::SemanticTokenType {
         match self {
+            SemanticTokenType::Module => lsp_types::SemanticTokenType::NAMESPACE,
             SemanticTokenType::Parameter => lsp_types::SemanticTokenType::PARAMETER,
             SemanticTokenType::Variable => lsp_types::SemanticTokenType::VARIABLE,
             SemanticTokenType::Symbol => lsp_types::SemanticTokenType::ENUM_MEMBER,
